+++ timeseq.go
@@ -1,7 +1,6 @@
 package timeseq
 
 import (
-	"encoding/binary"
 	"sort"
 	"time"
 )
@@ -22,6 +21,7 @@ type sortable struct {
 	Interface
 }
 
+// Less implements sort.Interface.Less()
 func (s sortable) Less(i, j int) bool {
 	return s.Time(i).Before(s.Time(j))
 }
@@ -31,10 +31,12 @@ func Sort(slice Interface) {
 	sort.Stable(sortable{Interface: slice})
 }
 
+// IsSorted reports whether data is sorted.
 func IsSorted(slice Interface) bool {
 	return sort.IsSorted(sortable{Interface: slice})
 }
 
+// Range return a sub slice of given sorted slice according to the interval
 func Range(slice Interface, interval Interval) Interface {
 	i := 0
 	if interval.NotBefore != nil {
@@ -52,135 +54,4 @@ func Range(slice Interface, interval Interval) Interface {
 		}
 	}
 	return slice.Slice(i, j)
-}
-
-// Interval indicates a continuous time range
-type Interval struct {
-	NotBefore *time.Time
-	NotAfter  *time.Time
-}
-
-// Contain return if time t is in the interval
-func (i Interval) Contain(t time.Time) bool {
-	if i.NotAfter != nil && t.After(*i.NotAfter) {
-		return false
-	}
-	if i.NotBefore != nil && t.Before(*i.NotBefore) {
-		return false
-	}
-	return true
-}
-
-// String returns the interval formatted using the RFC3339 format string
-func (i *Interval) String() string {
-	return i.Format(time.RFC3339)
-}
-
-// Format returns a textual representation of the time value formatted according to layout
-func (i *Interval) Format(layout string) string {
-	notBefore, notAfter := "nil", "nil"
-	if i.NotBefore != nil {
-		notBefore = i.NotBefore.Format(layout)
-	}
-	if i.NotAfter != nil {
-		notAfter = i.NotAfter.Format(layout)
-	}
-	return notBefore + "~" + notAfter
-}
-
-// BeginAt is alias of AfterOrEqual
-func (i Interval) BeginAt(t time.Time) Interval {
-	return i.AfterOrEqual(t)
-}
-
-// EndAt is alias of BeforeOrEqual
-func (i Interval) EndAt(t time.Time) Interval {
-	return i.BeforeOrEqual(t)
-}
-
-// BeforeOrEqual return a new Interval which not before t
-func (i Interval) BeforeOrEqual(t time.Time) Interval {
-	return Interval{
-		NotBefore: i.NotBefore,
-		NotAfter:  &t,
-	}
-}
-
-// AfterOrEqual return a new Interval which not after t
-func (i Interval) AfterOrEqual(t time.Time) Interval {
-	return Interval{
-		NotBefore: &t,
-		NotAfter:  i.NotAfter,
-	}
-}
-
-// Before return a new Interval which before t
-func (i Interval) Before(t time.Time) Interval {
-	t = t.Add(-1)
-	return Interval{
-		NotBefore: i.NotBefore,
-		NotAfter:  &t,
-	}
-}
-
-// After return a new Interval which after t
-func (i Interval) After(t time.Time) Interval {
-	t = t.Add(1)
-	return Interval{
-		NotBefore: &t,
-		NotAfter:  i.NotAfter,
-	}
-}
-
-// BeginAt is alias of AfterOrEqual
-func BeginAt(t time.Time) Interval {
-	return AfterOrEqual(t)
-}
-
-// EndAt is alias of BeforeOrEqual
-func EndAt(t time.Time) Interval {
-	return BeforeOrEqual(t)
-}
-
-// BeforeOrEqual return a new Interval which not before t
-func BeforeOrEqual(t time.Time) Interval {
-	return Interval{
-		NotAfter: &t,
-	}
-}
-
-// AfterOrEqual return a new Interval which not after t
-func AfterOrEqual(t time.Time) Interval {
-	return Interval{
-		NotBefore: &t,
-	}
-}
-
-// Before return a new Interval which before t
-func Before(t time.Time) Interval {
-	t = t.Add(-1)
-	return Interval{
-		NotAfter: &t,
-	}
-}
-
-// After return a new Interval which after t
-func After(t time.Time) Interval {
-	t = t.Add(1)
-	return Interval{
-		NotBefore: &t,
-	}
-}
-
-type timeKey [12]byte
-
-func (k timeKey) Time() time.Time {
-	return time.Unix(int64(binary.BigEndian.Uint64(k[:8])), int64(binary.BigEndian.Uint32(k[8:])))
-}
-
-func newTimeKey(t time.Time) timeKey {
-	var ret [12]byte
-	binary.BigEndian.PutUint64(ret[:8], uint64(t.Unix()))
-	binary.BigEndian.PutUint32(ret[8:], uint32(t.Nanosecond()))
-	return ret
 }
